requirements.txt
Jinja2==3.1.*
more-itertools==11.0.*
livereload==2.7.*
python-dotenv==1.2.1
environs==14.5.0
